requirements.txt
cv2-python
numpy